tutorial/requirements/requirements.txt
Django==3.0.7
djangorestframework==3.11.0
Pygments==2.6.1